external/unindp/requirements.txt
numpy==2.1.3
PyYAML==6.0.2
tqdm==4.66.4
xlrd==2.0.1
xlsxwriter==3.2.0
